AzDevOpsFromNuGetFeedUrlParser: parsing logic for NuGet Feed URLs
